Componentização do highcharts
